requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlmodel==0.0.24
psycopg[binary]==3.2.9
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
python-dotenv==1.1.1
email-validator==2.2.0
